Move functionalities of `MethodUtil` to `PropertyUtil`
After the debate of pr #426,
We decided to move all the methods of `MethodUtil`
to `PropertyUtil`.
There was a confusion of understanding
the name of the class `PropertyUtil`.
First I thought it was kind of a technical name
indicating that `PropertyUtil` deals with
java beans property.
But actually it wasn't just a technical name.
`PropertyUtil` means a tool to find out some
properties in the target class.
So to say, it is like a `PropertyFindingUtil`.
We don't have to change the utility name
even if it finds some methods of the target class,
not properties of it,
as it does what it is meant to do.
