+++ hamcrest/src/main/java/org/hamcrest/beans/PropertyUtil.java
@@ -2,6 +2,7 @@ package org.hamcrest.beans;
 
 import java.beans.IntrospectionException;
 import java.beans.Introspector;
+import java.beans.MethodDescriptor;
 import java.beans.PropertyDescriptor;
 
 /**
@@ -11,6 +12,7 @@ import java.beans.PropertyDescriptor;
  *
  * @author Iain McGinniss
  * @author Steve Freeman
+ * @author Uno Kim
  * @since 1.1.0
  */
 public class PropertyUtil {
@@ -53,6 +55,46 @@ public class PropertyUtil {
       } catch (IntrospectionException e) {
         throw new IllegalArgumentException("Could not get property descriptors for " + fromObj.getClass(), e);
       }
+    }
+
+    /**
+     * Returns the description of the method with the provided
+     * name on the provided object's interface.
+     * This is what you need when you try to find a property from a target object
+     * when it doesn't follow standard JavaBean specification, a Java Record for example.
+     *
+     * @param propertyName the object property name.
+     * @param fromObj the object to check.
+     * @param isNonVoid whether the method is non-void, which means the method has a return value.
+     * @return the descriptor of the method, or null if the method does not exist.
+     * @throws IllegalArgumentException if there's an introspection failure
+     * @see <a href="https://docs.oracle.com/en/java/javase/17/language/records.html">Java Records</a>
+     *
+     */
+    public static MethodDescriptor getMethodDescriptor(String propertyName, Object fromObj, boolean isNonVoid) throws IllegalArgumentException {
+        for (MethodDescriptor method : methodDescriptorsFor(fromObj, null)) {
+            if (method.getName().equals(propertyName) && (!isNonVoid || method.getMethod().getReturnType() != void.class)) {
+                return method;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Returns all the method descriptors for the class associated with the given object
+     *
+     * @param fromObj Use the class of this object
+     * @param stopClass Don't include any properties from this ancestor class upwards.
+     * @return Method descriptors
+     * @throws IllegalArgumentException if there's an introspection failure
+     */
+    public static MethodDescriptor[] methodDescriptorsFor(Object fromObj, Class<Object> stopClass) throws IllegalArgumentException {
+        try {
+            return Introspector.getBeanInfo(fromObj.getClass(), stopClass).getMethodDescriptors();
+        } catch (IntrospectionException e) {
+            throw new IllegalArgumentException("Could not get method descriptors for " + fromObj.getClass(), e);
+        }
     }
 
     /**
